+++ keycodes.go
@@ -1,5 +1,3 @@
-// +build !windows
-
 package promptui
 
 import "github.com/chzyer/readline"
@@ -10,8 +8,8 @@ var (
 	// KeyEnter is the default key for submission/selection.
 	KeyEnter rune = readline.CharEnter
 
-	// KeyBackspace is the default key for deleting input text.
-	KeyBackspace rune = readline.CharBackspace
+	// KeyCtrlH is the key for deleting input text.
+	KeyCtrlH rune = readline.CharCtrlH
 
 	// KeyPrev is the default key to go up during selection.
 	KeyPrev        rune = readline.CharPrev